+++ CHANGELOG.md
@@ -4,6 +4,22 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [2.58.0] - 2020-03-11
+
+Feature:
+- Add `showOtherPages` and `pagesClassName` to `Carousel`.
+
+Fix:
+- Remove `isActive` styles to `EngagementCard`.
+
+## [2.57.0] - 2020-03-06
+
+Feature:
+- Add `margin` utility classes for new steps and half steps.
+
+Fix:
+- Update `EngagementCard` styles.
+
 ## [2.56.0] - 2020-03-03
 
 Features: